Changes to make bulk inference easier: - load in features if already computed - allow computing features for multiple fastas in parallel - allow disabling Amber relaxation, b/c our sequences have X's

+++ run_alphafold.py
@@ -21,6 +21,9 @@ import random
 import sys
 import time
 from typing import Dict
+from concurrent.futures import ThreadPoolExecutor
+from functools import partial
+from tqdm.auto import tqdm
 
 from absl import app
 from absl import flags
@@ -88,6 +91,13 @@ flags.DEFINE_integer('random_seed', None, 'The random seed for the data '
                      'that even if this is set, Alphafold may still not be '
                      'deterministic, because processes like GPU inference are '
                      'nondeterministic.')
+flags.DEFINE_boolean('features_only', False, 'Only search MSAs/templates and dump features. '
+                     'Useful when running feature generation in parallel before doing GPU '
+                     'inference in serial')
+flags.DEFINE_integer('num_workers', 1, 'Number of parallel runs of feature generation. '
+                     'Only applies if features_only')
+flags.DEFINE_boolean('do_relax', True, 'Whether to run Amber relaxation. '
+                     'If sequence contains Xs, Amber will fail.')
 FLAGS = flags.FLAGS
 
 MAX_TEMPLATE_HITS = 20
@@ -112,7 +122,8 @@ def predict_structure(
     model_runners: Dict[str, model.RunModel],
     amber_relaxer: relax.AmberRelaxation,
     benchmark: bool,
-    random_seed: int):
+    random_seed: int,
+    features_only: bool):
   """Predicts structure using AlphaFold for the given sequence."""
   timings = {}
   output_dir = os.path.join(output_dir_base, fasta_name)
@@ -123,18 +134,27 @@ def predict_structure(
     os.makedirs(msa_output_dir)
 
   # Get features.
-  t_0 = time.time()
-  feature_dict = data_pipeline.process(
-      input_fasta_path=fasta_path,
-      msa_output_dir=msa_output_dir)
-  timings['features'] = time.time() - t_0
-
-  # Write out features as a pickled dictionary.
   features_output_path = os.path.join(output_dir, 'features.pkl')
-  with open(features_output_path, 'wb') as f:
-    pickle.dump(feature_dict, f, protocol=4)
+  if os.path.exists(features_output_path):
+    feature_dict = pickle.load(open(features_output_path, 'rb'))
+    logging.info('Loaded features for %s from %s', fasta_name, features_output_path)
+  else:
+    t_0 = time.time()
+    feature_dict = data_pipeline.process(
+        input_fasta_path=fasta_path,
+        msa_output_dir=msa_output_dir)
+    timings['features'] = time.time() - t_0
 
-  relaxed_pdbs = {}
+    # Write out features as a pickled dictionary.
+    with open(features_output_path, 'wb') as f:
+      pickle.dump(feature_dict, f, protocol=4)
+    logging.info('Generated features for %s, saved in %s', fasta_name, features_output_path)
+
+  if features_only:
+    logging.info('features_only set to True, so stopping here.')
+    return
+
+  unrelaxed_pdbs, relaxed_pdbs = {}, {}  # PDB file content strings
   plddts = {}
 
   # Run the models.
@@ -177,29 +197,31 @@ def predict_structure(
         b_factors=plddt_b_factors)
 
     unrelaxed_pdb_path = os.path.join(output_dir, f'unrelaxed_{model_name}.pdb')
+    unrelaxed_pdbs[model_name] = protein.to_pdb(unrelaxed_protein)
     with open(unrelaxed_pdb_path, 'w') as f:
-      f.write(protein.to_pdb(unrelaxed_protein))
+      f.write(unrelaxed_pdbs[model_name])
 
-    # Relax the prediction.
-    t_0 = time.time()
-    relaxed_pdb_str, _, _ = amber_relaxer.process(prot=unrelaxed_protein)
-    timings[f'relax_{model_name}'] = time.time() - t_0
+    if amber_relaxer:
+      # Relax the prediction.
+      t_0 = time.time()
+      relaxed_pdb_str, _, _ = amber_relaxer.process(prot=unrelaxed_protein)
+      timings[f'relax_{model_name}'] = time.time() - t_0
 
-    relaxed_pdbs[model_name] = relaxed_pdb_str
+      relaxed_pdbs[model_name] = relaxed_pdb_str
 
-    # Save the relaxed PDB.
-    relaxed_output_path = os.path.join(output_dir, f'relaxed_{model_name}.pdb')
-    with open(relaxed_output_path, 'w') as f:
-      f.write(relaxed_pdb_str)
+      # Save the relaxed PDB.
+      relaxed_output_path = os.path.join(output_dir, f'relaxed_{model_name}.pdb')
+      with open(relaxed_output_path, 'w') as f:
+        f.write(relaxed_pdb_str)
 
-  # Rank by pLDDT and write out relaxed PDBs in rank order.
+  # Rank by pLDDT and write out PDBs (relaxed if available, else unrelaxed) in rank order.
   ranked_order = []
   for idx, (model_name, _) in enumerate(
       sorted(plddts.items(), key=lambda x: x[1], reverse=True)):
     ranked_order.append(model_name)
     ranked_output_path = os.path.join(output_dir, f'ranked_{idx}.pdb')
     with open(ranked_output_path, 'w') as f:
-      f.write(relaxed_pdbs[model_name])
+      f.write((relaxed_pdbs if amber_relaxer else unrelaxed_pdbs)[model_name])
 
   ranking_output_path = os.path.join(output_dir, 'ranking_debug.json')
   with open(ranking_output_path, 'w') as f:
@@ -223,6 +245,8 @@ def main(argv):
               should_be_set=not use_small_bfd)
   _check_flag('uniclust30_database_path', FLAGS.preset,
               should_be_set=not use_small_bfd)
+  if FLAGS.num_workers > 1 and not FLAGS.features_only:
+    raise app.UsageError('num_workers>1 only applies to features_only')
 
   if FLAGS.preset in ('reduced_dbs', 'full_dbs'):
     num_ensemble = 1
@@ -280,16 +304,29 @@ def main(argv):
   logging.info('Using random seed %d for the data pipeline', random_seed)
 
   # Predict structure for each of the sequences.
-  for fasta_path, fasta_name in zip(FLAGS.fasta_paths, fasta_names):
-    predict_structure(
-        fasta_path=fasta_path,
-        fasta_name=fasta_name,
-        output_dir_base=FLAGS.output_dir,
-        data_pipeline=data_pipeline,
-        model_runners=model_runners,
-        amber_relaxer=amber_relaxer,
-        benchmark=FLAGS.benchmark,
-        random_seed=random_seed)
+  inputs = list(zip(FLAGS.fasta_paths, fasta_names))
+
+  def fn(inp):
+    fasta_path, fasta_name = inp
+    return predict_structure(
+      fasta_path=fasta_path,
+      fasta_name=fasta_name,
+      output_dir_base=FLAGS.output_dir,
+      data_pipeline=data_pipeline,
+      model_runners=model_runners,
+      amber_relaxer=amber_relaxer if FLAGS.do_relax else None,
+      benchmark=FLAGS.benchmark,
+      random_seed=random_seed,
+      features_only=FLAGS.features_only)
+
+  if FLAGS.num_workers > 1:
+    # use thread pool rather than process pool b/c the work is done in subprocesses anyway
+    logging.info('Processing %d inputs in parallel using %d workers', len(inputs), FLAGS.num_workers)
+    with ThreadPoolExecutor(max_workers=FLAGS.num_workers) as executor:
+      ret = list(tqdm( executor.map(fn, inputs), total=len(inputs) ))
+  else:
+    logging.info('Processing %d inputs serially', len(inputs))
+    ret = list(tqdm( map(fn, inputs), total=len(inputs) ))
 
 
 if __name__ == '__main__':
@@ -305,6 +342,8 @@ if __name__ == '__main__':
       'template_mmcif_dir',
       'max_template_date',
       'obsolete_pdbs_path',
+      'features_only',
+      'do_relax',
   ])
 
   app.run(main)
